refactor: extract shared @chroxy/protocol/project module (audit P2-2, closes #5850)
The project-name derivation was duplicated across the claude-hooks emitter
(project.js) and the server's POST /api/events fallback (event-ingest.js
deriveProjectFromCwd). The hook half had accreted the chroxy-worktree fixes
(#5439/#5464/#5483) the server half lacked, so every fix was re-derived. Unify
both behind one source of truth.

- New packages/protocol/src/project.ts — a Zod-free module (node builtins only)
  exporting deriveProject + classifyNonProjectCwd + worktreeParent (hook surface)
  and deriveProjectFromCwd (server surface), with the shared chroxy-worktree
  helpers. Cross-platform paths (isAbsolute/relative/sep); honors BOTH former
  copies' test-override env names so neither suite's fixtures change.
- protocol exposes it via a `./project` subpath export → dist/project.js (kept
  out of the Zod barrel; subpath imports in ~7ms, well under the hook's 100ms
  budget). Enabled `@types/node` for the package typecheck (node-only subpath).
- claude-hooks/src/project.js is now a thin re-export; @chroxy/protocol promoted
  from devDependency to a real (workspace) dependency. Distribution is monorepo-
  only (README), so workspace resolution covers it; a package.json note flags the
  bundle-if-ever-published-standalone caveat.
- event-ingest.js imports deriveProjectFromCwd from the shared module and drops
  its local copy + now-unused imports.

Design A (real dep) over bundling: chroxy is not published to npm; the hook always
runs from an installed workspace where @chroxy/protocol resolves (private only
blocks publishing that package, not workspace deps). True single-source DRY with
no drift-prone committed copy.
