+++ packages/protocol/package.json
@@ -14,6 +14,10 @@
     "./schemas": {
       "import": "./dist/schemas/index.js",
       "types": "./dist/schemas/index.d.ts"
+    },
+    "./project": {
+      "import": "./dist/project.js",
+      "types": "./dist/project.d.ts"
     }
   },
   "scripts": {
@@ -25,6 +29,7 @@
     "zod": "^4.3.6"
   },
   "devDependencies": {
+    "@types/node": "*",
     "tsx": "^4.19.2",
     "typescript": "^5.6.0"
   }
